refactor: move class PullCardMessage and TaxPaymentMessage to message package and implements GameMessage Interface
