--- app/src/main/java/at/aau/serg/websocketbrokerdemo/data/PullCardMessage.kt
@@ -1,7 +1,0 @@
-package at.aau.serg.websocketbrokerdemo.data
-
-data class PullCardMessage(
-    val type: String = "PULL_CARD",
-    val playerId: String,
-    val cardType: String
-)
