+++ app/src/main/java/at/aau/serg/websocketbrokerdemo/data/messages/PullCardMessage.kt
@@ -1,0 +1,7 @@
+package at.aau.serg.websocketbrokerdemo.data.messages
+
+data class PullCardMessage(
+   override val type: String = "PULL_CARD",
+    val playerId: String,
+    val cardType: String
+):GameMessage
