+++ app/src/main/java/at/aau/serg/websocketbrokerdemo/data/messages/TaxPaymentMessage.kt
@@ -1,8 +1,8 @@
-package at.aau.serg.websocketbrokerdemo.data
+package at.aau.serg.websocketbrokerdemo.data.messages
 
 data class TaxPaymentMessage(
-    val type: String = "TAX_PAYMENT",
+  override  val type: String = "TAX_PAYMENT",
     val playerId: String,
     val amount: Int,
     val taxType: String  // "EINKOMMENSTEUER" or "ZUSATZSTEUER"
-)
+):GameMessage
